% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.hqreg.R
\name{coef.cv.hqreg}
\alias{coef.cv.hqreg}
\title{coefficient of cross validated models}
\usage{
\method{coef}{cv.hqreg}(object, lambda = c("lambda.1se", "lambda.min"),
  ...)
}
\arguments{
\item{object}{cross validated hqreg model}

\item{lambda}{choose "lambda.1se" or "lambda.min"}
}
\description{
coefficient of cross validated models
}
