% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.hqreg.R
\name{coef.hqreg}
\alias{coef.hqreg}
\title{coefficient of cross validated models}
\usage{
\method{coef}{hqreg}(object, lambda, exact = FALSE, ...)
}
\arguments{
\item{object}{cross validated hqreg model}

\item{lambda}{a vector of lambdas}
}
\description{
coefficient of cross validated models
}
