% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.hqreg.R
\name{cv.hqreg}
\alias{cv.hqreg}
\title{cross validate huber/quantile regression}
\usage{
cv.hqreg(X, y, ..., FUN = c("hqreg", "hqreg_raw"), parallel = F,
  nfolds = 10, fold.id, type.measure = c("deviance", "mse", "mae"),
  seed)
}
\arguments{
\item{y}{response}

\item{FUN}{use hqreg or hqreg_raw function}

\item{parallel}{use parallel computing or not}

\item{nfolds}{number of folds}

\item{fold.id}{id of folds}

\item{type.measure}{the methods to measure the distance}

\item{x}{predictors}
}
\description{
cross validate huber/quantile regression
}
