% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.hqreg.R
\name{cvf}
\alias{cvf}
\title{cross validation function}
\usage{
cvf(i, XX, y, weights, fold.id, cv.args, measure.args, FUN)
}
\arguments{
\item{i}{ith fold}

\item{y}{response}

\item{weights}{weights of observations}

\item{fold.id}{id of folds}

\item{cv.args}{cross validation args}

\item{measure.args}{measurement arguments}

\item{FUN}{function of use: hqreg or hqreg_raw}

\item{xx}{predictors}
}
\description{
cross validation function
}
