% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hqreg_raw.R
\name{hqreg_raw}
\alias{hqreg_raw}
\title{high dimensional quantile regression with raw data}
\usage{
hqreg_raw(X, y, weights = NULL, method = c("huber", "quantile", "ls"),
  gamma = IQR(y)/10, tau = 0.5, alpha = 1, nlambda = 100,
  lambda.min = 0.05, lambda, intercept = TRUE, screen = c("ASR",
  "SR", "none"), max.iter = 10000, eps = 1e-07, dfmax = ncol(X) + 1,
  penalty.factor = rep(1, ncol(X)), message = FALSE)
}
\arguments{
\item{y}{response}

\item{weights}{weights of observations}

\item{method}{type of loss function}

\item{gamma}{gamma used for huber loss}

\item{tau}{tau used for quantile regression}

\item{alpha}{mixing parameter of elastic net}

\item{penalty.factor}{numbers to penalise each predictor}

\item{x}{predictors}
}
\description{
high dimensional quantile regression with raw data
}
