% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.hqreg.R
\name{measure.hqreg}
\alias{measure.hqreg}
\title{measure the function loss}
\usage{
measure.hqreg(r, weights = rep(1, length(r)), args)
}
\arguments{
\item{r}{residual}

\item{weights}{weights of observations}

\item{args}{arguments of type measure}
}
\value{
value of loss function
}
\description{
measure the function loss
}
