% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.hqreg.R
\name{plot.cv.hqreg}
\alias{plot.cv.hqreg}
\title{plot cross validated hqreg model}
\usage{
\method{plot}{cv.hqreg}(x, log.l = TRUE, nvars = TRUE, ...)
}
\arguments{
\item{x}{cross validated hqreg model}
}
\description{
plot cross validated hqreg model
}
