% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hqreg.R
\name{plot.hqreg}
\alias{plot.hqreg}
\title{plot of hqreg model}
\usage{
\method{plot}{hqreg}(x, xvar = c("lambda", "norm"), log.l = TRUE,
  nvars = TRUE, alpha = 1, ...)
}
\arguments{
\item{x}{hqreg model object}
}
\description{
plot of hqreg model
}
