% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.hqreg.R
\name{predict.cv.hqreg}
\alias{predict.cv.hqreg}
\title{predict model by choosing the cross validated hqreg model}
\usage{
\method{predict}{cv.hqreg}(object, X, lambda = c("lambda.1se",
  "lambda.min"), type = c("response", "coefficients", "nvars"), ...)
}
\arguments{
\item{object}{cross validated hqreg models}

\item{x}{new data}
}
\description{
predict model by choosing the cross validated hqreg model
}
