% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.hqreg.R
\name{predict.hqreg}
\alias{predict.hqreg}
\title{predict model hqreg model}
\usage{
\method{predict}{hqreg}(object, X, lambda, type = c("response",
  "coefficients", "nvars"), exact = FALSE, ...)
}
\arguments{
\item{object}{cross validated hqreg models}

\item{lambda}{predict at which lambda}

\item{x}{new data}
}
\description{
predict model hqreg model
}
