% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.standardize.R
\name{weighted.standardise}
\alias{weighted.standardise}
\title{weighted standardise the data}
\usage{
weighted.standardise(x, weights)
}
\arguments{
\item{x}{observations}

\item{weights}{weights of observations}
}
\description{
weighted standardise the data
}
